refactor(lex-cli): split main.rs and transforms.rs into cohesive submodules
